+++ zakozhurnikova.kristina/P3/removeSpaces.h
@@ -1,0 +1,7 @@
+#ifndef REMOVESPACES_H
+#define REMOVESPACES_H
+#include <cstddef>
+
+void removeSpaces(char *dest, const char *src, size_t buff);
+
+#endif
